packages/mcp-server/pyproject.toml
[tool.poetry]
name = "mcp-bridge-server"
version = "0.1.0" # Keep versions synced or manage independently
description = "MCP Bridge Server exposing tools for AI agents."
authors = ["Richard Chukwu <richinex@gmail.com>"]
license = "MIT" # Or your license
readme = "../../README.md"
packages = [{include = "mcp_server", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.10,<4.0"

# Core Server Dependencies (adjust based on actual server needs)
fastapi = "^0.115.12" # Match version used previously
uvicorn = "^0.34.1"  # Match version used previously
sse-starlette = "^2.2.1" # Match version used previously
docker = "^7.1.0" # Match version used previously
pydantic = "^2.5.2" # Likely needed for models
python-dotenv = "^1.0.0" # If server uses .env
aioredis = "^2.0.1"
redis = "^5.2.1"
docstring-parser = "^0.15"
networkx = "^2.8.8"
celery = "^5.5.1"
# Dependency for Stdio Transport (if server supports it)
mcp = { version = "^1.6.0" } # Match version used previously, likely needed for protocol implementation

# Dependencies for Tools (Add ALL dependencies needed by your tools)
selenium = "^4.9.0"
requests = "^2.28.1"
beautifulsoup4 = "^4.12.0"
feedparser = "^6.0.10"
PyPDF2 = "^3.0.1"
trafilatura = "^2.0.0"
lxml = "^5.4.0"
lxml-html-clean = "^0.4.2"
google-api-python-client = "^2.169.0"
google-auth-httplib2 = "^0.2.0"
google-auth-oauthlib = "^1.2.2"
jsonrpcserver = "^5.0.9"
jsonrpcclient = "^4.0.3"
pytz = "^2025.2"
pymupdf4llm = "^0.0.22"
textual = "^3.2.0"
GitPython = "^3.1.44"
asyncpraw = "<7.8.1"
# Add any others used in tools/ directory...

[tool.poetry.scripts]
# Assumes your server entry point (run.py) uses Typer or has a main()
# Adjust the entry point based on how run.py is structured
# Option 1: If run.py uses Typer with an app object
mcp-bridge-server = "mcp_server.run:app"
# Option 2: If run.py has a main() function to start the server
# mcp-bridge-server = "mcp_server.run:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"